requirements.txt
openai>=1.0.0
python-dotenv
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
pydantic[email]>=2.0.0
sqlalchemy>=2.0.0
docling>=2.0.0
